Move dict-related API to the correct header file.
